fix cookFoodLeaf `Production.interact()`
You need to interact with the finished item's ID/name, not the unfinished. So it seems good enough for now to convert id -> name and strip `Raw ` from the name. This will run into issues later where the uncooked equivalent doesn't have `Raw ` in its name but for the foreseeable future I don't see us cooking anything but fish so we'll go with this.